fixing bug related to epoch time (1970-01-01)
reference.xml had DATE_OF_BIRTH="1970-01-01" and the test had new
Date(0). However, if TimeZone is not set to UTC/GMT, then new Date(0)
will not result in "1970-01-01".

So, in reference.xml I replaced "1970-01-01" with a placeholder
"[EPOCH_TIME]" and in the ReplacementDataSet I added replacement from it
to new Date(0).