DataSetBuilder was ignoring caseSensitiveTableNames to build DataSet.
Unfortunately the class CachedDataSet provided by DBUnit only allows
users to inform "caseSensitiveTableNames" along with IDataSetProducer.

But the strategy used by DataSetBuilder doesn't require or support the
use of IDataSetProducer, so the only way I found to set
"caseSensitiveTableNames" to the CachedDataSet built by DataSetBuilder
was to force an injection through Reflections API (ugly, but necessary
until DBUnit provides a way to construct a CachedDataSet passing
"_caseSensitiveTableNames" and no IDataSetProducer).